requirements.txt
selenium==3.141.0
pyyaml==5.3.1